fix editor in sighting view
replace on directives tries to merge the replaced element with the
directive. When it finds two attributes with the same name it concats
the values with a space in between. This doesn't work at all with ng-if,
which is used by the sighting-view.. aaahhrrngular.
also fixes the width of the iframe